Refactored registration logic
