+++ app/.meteor/release
@@ -1,1 +1,1 @@
-METEOR@1.3.2.2
+METEOR@1.3.2.4
