+++ app/.meteor/versions
@@ -4,9 +4,12 @@ accounts-ui@1.1.9
 accounts-ui-unstyled@1.1.12
 alanning:roles@1.2.15
 aldeed:autoform@5.8.1
-aldeed:collection2@2.3.1
+aldeed:collection2@2.9.1
+aldeed:collection2-core@1.1.1
 aldeed:delete-button@2.0.0
-aldeed:simple-schema@1.1.0
+aldeed:schema-deny@1.0.1
+aldeed:schema-index@1.0.1
+aldeed:simple-schema@1.5.3
 allow-deny@1.0.4
 autoupdate@1.2.9
 babel-compiler@6.6.4
@@ -22,7 +25,7 @@ caching-html-compiler@1.0.6
 callback-hook@1.0.8
 check@1.2.1
 ddp@1.2.5
-ddp-client@1.2.6
+ddp-client@1.2.7
 ddp-common@1.2.5
 ddp-rate-limiter@1.0.4
 ddp-server@1.2.6
@@ -54,24 +57,26 @@ less@2.6.0
 livedata@1.0.18
 localstorage@1.0.9
 logging@1.0.12
-markdown@1.0.9
+mdg:validation-error@0.2.0
 meteor@1.1.14
 meteor-base@1.0.4
 minifier-css@1.1.11
 minifier-js@1.1.11
-minimongo@1.0.15
+minimongo@1.0.16
 mobile-experience@1.0.4
 mobile-status-bar@1.0.12
 modules@0.6.1
 modules-runtime@0.6.3
-momentjs:moment@2.10.6
+momentjs:moment@2.13.1
 mongo@1.1.7
 mongo-id@1.0.4
 npm-bcrypt@0.8.5
 npm-mongo@1.4.43
 observe-sequence@1.0.11
 ordered-dict@1.0.7
+perak:markdown@1.0.5
 promise@0.6.7
+raix:eventemitter@0.1.3
 random@1.0.9
 rate-limit@1.0.4
 reactive-dict@1.1.7
@@ -96,3 +101,4 @@ underscore@1.0.8
 url@1.0.9
 webapp@1.2.8
 webapp-hashing@1.0.9
+wtfzn:bootstrap-markdown@2.8.0
