--- app/client/main.html
@@ -1,11 +1,0 @@
-<head>
-  <title>simple</title>
-</head>
-
-<body>
-
-</body>
-
-<template name="hello">
-
-</template>
