+++ app/client/templates/admin/create_question/create_question.html
@@ -11,6 +11,5 @@
     {{else}}
       {{> NotFound}}
     {{/if}}
-
   </div>
 </template>
