+++ app/client/templates/admin/create_question/create_question.js
@@ -2,6 +2,7 @@
 /* CreateQuestion: Event Handlers */
 /*****************************************************************************/
 Template.CreateQuestion.events({
+  
 });
 
 /*****************************************************************************/
