+++ app/client/templates/admin/register/register.html
@@ -19,7 +19,7 @@
             <option value="admin">Admin</option>
           </select>
         </div>
-        <input class="btn btn-success" type="submit" value="Register">
+        <input class="btn btn-primary" type="submit" value="Register">
         <a href="/" class="btn btn-danger">Cancel</a>
       </form>
     {{else}}
