+++ app/client/templates/admin/register/register.js
@@ -8,35 +8,9 @@ Template.Register.events({
     var passwordVar = event.target.regPassword.value;
     var role = event.target.regRole.value;
 
-    var _createUserAsync = function(userVar, passwordVar) {
-      return new Promise((resolve, reject) => {
-        Accounts.createUser({
-          username: userVar,
-          password: passwordVar
-        },
-        function(error,result){
-          if(error){
-            console.log(error.reason);
-            reject(error.reason);
-          }else{
-            var userId = Meteor.userId();
-            resolve(userId);
-          }
-        })
+    Meteor.call('createGroup', userVar, passwordVar, role);
 
-      });
-    };
-
-    _createUserAsync(userVar, passwordVar)
-      .then(userId => {
-        if(role === 'admin'){
-          Meteor.call('adminRole', userId);
-        }else{
-          Meteor.call('defaultRole', userId);
-        }
-        Router.go('register');
-      });
-
+    event.target.reset();
     /*if(role === 'admin'){
       Roles.setUserRoles(userId, 'admin');
     }else{
